+++ src/Ieemdb.Web/ClientApp/src/Styles/Login.css
@@ -1,5 +1,6 @@
 .column-right-Login{
     background-color: #282828;
+    color: white;
     height: 100vh;
     padding: 0%;
     width: 80%;
